reorganize core extensions
