+++ card/config/initializers/01_core_extensions/array.rb
@@ -1,0 +1,9 @@
+module CoreExtensions
+  module Array
+    def to_pointer_content
+      map do |item|
+        "[[#{item}]]"
+      end.join "\n"
+    end
+  end
+end
