--- card/config/initializers/01_init_ruby_extensions.rb
@@ -1,3 +1,0 @@
-# -*- encoding : utf-8 -*-
-
-require 'card/core_ext'
