+++ card/config/initializers/extensions.rb
@@ -1,0 +1,1 @@
+Kaminari::Helpers::Tag.include Extensions::Kaminari::Helpers::Tag
